Move Tooltip to components/base and rename to one word
