+++ airbyte-webapp/src/components/base/Tooltip/Tooltip.module.scss
@@ -1,6 +1,6 @@
-@use "../../scss/colors";
-@use "../../scss/variables";
-@use "../../scss/z-indices";
+@use "../../../scss/colors";
+@use "../../../scss/variables";
+@use "../../../scss/z-indices";
 
 .container {
   display: inline;
