+++ airbyte-webapp/src/components/base/Tooltip/TooltipLearnMoreLink.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/variables";
+@use "../../../scss/variables";
 
 .container {
   margin-top: variables.$spacing-md;
