+++ airbyte-webapp/src/components/base/Tooltip/TooltipTable.module.scss
@@ -1,5 +1,5 @@
-@use "../../scss/colors";
-@use "../../scss/variables";
+@use "../../../scss/colors";
+@use "../../../scss/variables";
 
 .label {
   color: rgba(colors.$white, 0.7);
