+++ airbyte-webapp/src/components/base/Tooltip/index.ts
@@ -1,8 +1,4 @@
-import { ToolTip } from "./ToolTip";
-
 export * from "./context";
+export * from "./Tooltip";
 export * from "./TooltipTable";
 export * from "./types";
-
-export default ToolTip;
-export { ToolTip };
